backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
PyMySQL==1.1.1
openai>=1.0.0
mlx-lm>=0.18.0
